requirements.txt
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-rest-framework==0.1.0
pip-chill==1.0.3
psycopg2==2.9.6
python-dotenv==1.0.0
shroomdk==1.0.2
web3==6.2.0
